Add `this.super` to priviledged functions namespace
Before this commit it was not possible to call `this.super` from
priviledged functions. Thus it was not possible to correctly overwrite
a priviledged function.

Note that it is not possible execute a newly created priviledged
function with a reference to `this.super` from inside the `init` method.
The wrapping happens only after the `init` method has returned.