Allow init functions to make their own super() call.
Before this commit the super constructor with no args
would be called before every init() call. And it was
not possible to do it with arguments, because this.super
would not behave as expected.

Now this.super() behaves as expected, and the user
can access it when needed. If the user does not
call this.super() inside init, this call will be done
automatically with no args before init is executed.

If the user want to prevent this.super() from being
executed he just needs to execute a line like this:

  `this.super = false`