Add reinit method, used when init is called 2 times
Consider the instance b1 of class b which is subclass of a.

Before reinit, the call b1.apply(a, 'arguments'), would do nothing
and the arguments would be ignored, leaving the user with no feedback.

Now the user can define `this.reinit` so it can receive the arguments
and produce the expected behavior.

Another name I considered to `reinit` was `post_init` because
what it is supposed to do is the same thing as init, except that it
would not recreate the already instantiated attributes.

Reinit is useful when the final user does not know that b inherits from a.

In this situation he may belive that these two code samples
would do the same thing:

1. `(new Class()).apply(a, arg).apply(b, arg)`

2. `(new Class()).apply(b, arg).apply(a, arg)`

When in fact they would not because when b is initialized first
a is also initialized, causing the second `apply(a, arg)` to be ignored.

Now thanks to reinit, the ignoring behavior can be fixed,
and the final user does not need to know all the class hierarchy.