Fix issue #2 - Add this.overwrite function
Issue 2 is caused by incomplete function overwritting,
inside init. To solve this problem this.overwrite
was added to init namespace.

This function should be called to create priviledged
function from inside init. If the user do not use it,
the class will probably still work but it is safer
to always declare priviledged function with it.

Notes:
 - `this.overwrite` only exists during the execution of init.
 - Added new tests to `class.js`