Fix bug when a super call was falsely detected on init
Before this commit if a user overwrites a privileged function,
from `init`, and inside this function he calls `this.super()`
the `Class` framework would believe that `init` has called
`this.super()`, so it would not automatically call it for him.

This commit adds a better `RegExp` detector to make sure it will
detect `this.super()` calls correctly even when there are internal
functions calling `this.super()`.